# Performance: memomize LinesClassifier::no_cov_line