@BMorearty 2018-03-09
P.S. Those benchmarks were in Ruby 2.1.8