@BMorearty 2018-03-09
Please hold off on merging this until I resolve an issue I'm getting in our CI, although I can't repro it locally:

```
.../vendor/bundle/ruby/2.1.0/gems/simplecov-0.15.1.abnb/lib/simplecov/lines_classifier.rb:23:in `=~': invalid byte sequence in UTF-8 (ArgumentError)
```