@BMorearty 2018-03-10
Update: feel free to merge this. The issue I'm seeing is unrelated to this change. I get the same behavior when I upgrade from 0.14.1 to 0.15.1. Haven't yet figured out why.